Added a quickcheck test for stripTrailingDigits
The coverage report showed that the existing tests didn't cover one of
the paths, so I made a simple property test to cover the rest.

This is slightly awkward in go, but the bulk of the test is generating
random input with some number of digits on the left, then some number
of digits on the right of a decimal point.  If there are no digits on
the right, it doesn't add the decimal point.  Then it emits a
reasonably ranged digit input.